backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
boto3==1.29.7
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
drf-spectacular==0.26.5
Pillow==10.1.0
paypalrestsdk==1.7.1
